bot/requirements.txt
python-telegram-bot
aiogram
aiosqlite
apscheduler